client-server/server/requirements.txt
azure-ai-language-conversations==1.1.0
azure-ai-textanalytics==5.3.0
azure-ai-translation-text==1.0.0
azure-cognitiveservices-speech==1.37.0
azure-core==1.30.2
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
langchain==0.2.3
langchain-openai==0.1.8
openai==1.32.0
pyaudio==0.2.14
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
requests==2.32.3